Show feedback on certain actions
Show feedback for actions that don't navigate to different views.
